+++ RELEASENOTES.md
@@ -27,6 +27,8 @@
     *   Make `AudioTrackBufferSizeProvider` a public interface.
     *   Add `WrappingMediaSource` to simplify wrapping a single `MediaSource`
         ([#7279](https://github.com/google/ExoPlayer/issues/7279)).
+    *   Add `ExoPlayer.setPreferredAudioDevice` to set the preferred audio
+        output device ([#135](https://github.com/androidx/media/issues/135)).
 *   Metadata:
     *   `MetadataRenderer` can now be configured to render metadata as soon as
         they are available. Create an instance with